rc2/requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.1
comm==0.1.2
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
flatbuffers==23.1.4
fonttools==4.41.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==7.4.9
jupyter_core==5.1.3
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mlxtend==0.22.0
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.25.1
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.0.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.11.1
seaborn==0.13.0
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
torch==2.0.0+cpu
torch-tb-profiler==0.4.1
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tornado==6.2
traitlets==5.8.1
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
